chore: move AssetOnLedger into account management domain
